Add a new function: applyHoverHightlight()
This checks your cursor everytime is moves. If your cursor is on a
link/anchor, it will highlight that link. The new highlight group
"w3mLinkHover" can be used to format the colorchange (SpecialKey by
default). The highlighting can be turned on or off by changing.
g:w3m#set_hover_on (val > 0 --> ON, else OFF). Hover highlighting is
turned on by default.