Change b:link_index_list into b:anchor_list
b:anchor_list is capable of much more and is therefore much more powerful.
It is a list of line-associated lists like b:link_index_list, but instead of
tag-indexes ( --> b:tag_list), they are complete anchors.

Right now, each anchor is a dictionary with "attr", "startCol", "endCol", and "line".

    " example usage
    " the first anchor on line 55
    let link = b:anchor_index[55][1]
    " the first column of the link
    let s = link.startCol
    " the final column
    let e = link.endCol
    " the line
    let l = link.line
    " the link-reference
    let r = link.attr.href